S6: DispatchLaunchTest moves to its own module
tests/test_dispatch_launch.py was 686 physical lines, the one file
check_source_sizes still named after S3's split. S4 has landed, so the
seam S3 left alone is free: DispatchLaunchTest exercises the `dispatch`
command end to end, where the two classes that stay exercise launch
resolution and `land`.

The class moves byte-identical (lines 162-460 at the parent, diffed
against `git show HEAD:` before anything ran) under a header carrying
only what it reads; the kept file drops `import threading`, the one
import nothing left in it uses. 686 -> 384 and 340. warnings 37 -> 36.

Manifest regenerated: the five DispatchLaunchTest mutation owners
re-key to the new module and keep their selected-module-boundary
ruling; discovery 2156 and owners 384 unchanged either side.
