+++ tests/serial_compat_manifest.json
@@ -103,16 +103,6 @@
    "test_constitution.WriteTimePointerTest.test_documentation_lists_it_among_the_day_zero_artifacts",
    "test_cutcheck.CutcheckTest.test_closed_graph_passes_even_when_candidates_may_overlap",
    "test_cutcheck.CutcheckTest.test_dangling_dependency_and_cycle_fail",
-   "test_dispatch_launch.DispatchLaunchTest.test_a_sealed_profile_override_resolves_the_planner_binding",
-   "test_dispatch_launch.DispatchLaunchTest.test_an_unknown_host_refuses_before_the_attempt_is_opened",
-   "test_dispatch_launch.DispatchLaunchTest.test_an_unresolved_role_refuses_before_the_attempt_is_opened",
-   "test_dispatch_launch.DispatchLaunchTest.test_dispatching_a_sealed_pending_item_readies_it_without_hanging",
-   "test_dispatch_launch.DispatchLaunchTest.test_no_lane_asks_a_child_for_a_verdict_token",
-   "test_dispatch_launch.DispatchLaunchTest.test_the_dispatch_carries_the_launch_its_host_record_declares",
-   "test_dispatch_launch.DispatchLaunchTest.test_the_host_flag_selects_another_hosts_binding",
-   "test_dispatch_launch.DispatchLaunchTest.test_the_prompt_carries_every_fact_a_child_cannot_derive",
-   "test_dispatch_launch.DispatchLaunchTest.test_the_prompt_names_the_identities_every_record_is_filed_under",
-   "test_dispatch_launch.DispatchLaunchTest.test_the_removed_carriage_selectors_are_ordinary_unknown_arguments",
    "test_dispatch_launch.LandTest.test_a_land_with_no_predicate_and_no_status_refuses_before_integrating",
    "test_dispatch_launch.LandTest.test_a_suspended_join_keeps_the_tree_its_handoff_resumes_in",
    "test_dispatch_launch.LandTest.test_land_imports_an_outcome_handed_to_it_on_standard_input",
@@ -130,6 +120,16 @@
    "test_dispatch_launch.LaunchResolutionTest.test_the_effort_is_the_one_this_host_spells_however_it_spells_it",
    "test_dispatch_launch.LaunchResolutionTest.test_the_profile_override_decides_the_role_the_skill_declared",
    "test_dispatch_launch.LaunchResolutionTest.test_the_selected_host_is_the_flag_then_the_environment_then_claude",
+   "test_dispatch_launch_command.DispatchLaunchTest.test_a_sealed_profile_override_resolves_the_planner_binding",
+   "test_dispatch_launch_command.DispatchLaunchTest.test_an_unknown_host_refuses_before_the_attempt_is_opened",
+   "test_dispatch_launch_command.DispatchLaunchTest.test_an_unresolved_role_refuses_before_the_attempt_is_opened",
+   "test_dispatch_launch_command.DispatchLaunchTest.test_dispatching_a_sealed_pending_item_readies_it_without_hanging",
+   "test_dispatch_launch_command.DispatchLaunchTest.test_no_lane_asks_a_child_for_a_verdict_token",
+   "test_dispatch_launch_command.DispatchLaunchTest.test_the_dispatch_carries_the_launch_its_host_record_declares",
+   "test_dispatch_launch_command.DispatchLaunchTest.test_the_host_flag_selects_another_hosts_binding",
+   "test_dispatch_launch_command.DispatchLaunchTest.test_the_prompt_carries_every_fact_a_child_cannot_derive",
+   "test_dispatch_launch_command.DispatchLaunchTest.test_the_prompt_names_the_identities_every_record_is_filed_under",
+   "test_dispatch_launch_command.DispatchLaunchTest.test_the_removed_carriage_selectors_are_ordinary_unknown_arguments",
    "test_dispatch_launch_lines.IdentityLineTest.test_no_skill_path_drops_only_the_file_clause",
    "test_dispatch_launch_lines.IdentityLineTest.test_the_first_line_names_the_skill_tool_and_the_verbatim_argument",
    "test_dispatch_launch_lines.IdentityLineTest.test_the_first_line_tells_a_fork_already_inside_to_work_in_place",
@@ -2159,7 +2159,7 @@
    "tests.test_workspace_cases.start_cases.TestTheStampPreservesTheTicketsByteDomain.test_only_the_two_stamped_lines_differ_from_the_prior_bytes",
    "tests.test_workspace_cases.start_cases.TestTicketsPayloadIsGradedNotItsExitStatus.test_an_error_payload_returned_at_exit_zero_is_a_failure"
   ],
-  "sha256": "347d242ecaf6e790fad2f7d1dfa504d38e77427e8c2dbb7748abdd51f14e4dcd"
+  "sha256": "1df1feba4fa51188f5f8071c13285c729fa894ef7c76d93cf350630b8e927cc5"
  },
  "mutation_owners": [
   {
@@ -2247,47 +2247,6 @@
   },
   {
    "module": "tests.test_dispatch_launch",
-   "owner": "DispatchLaunchTest.established",
-   "restoration": "selected-module-boundary",
-   "seams": [
-    "monkeypatch"
-   ]
-  },
-  {
-   "module": "tests.test_dispatch_launch",
-   "owner": "DispatchLaunchTest.setUp",
-   "restoration": "selected-module-boundary",
-   "seams": [
-    "environment",
-    "threads"
-   ]
-  },
-  {
-   "module": "tests.test_dispatch_launch",
-   "owner": "DispatchLaunchTest.test_an_unknown_host_refuses_before_the_attempt_is_opened",
-   "restoration": "selected-module-boundary",
-   "seams": [
-    "monkeypatch"
-   ]
-  },
-  {
-   "module": "tests.test_dispatch_launch",
-   "owner": "DispatchLaunchTest.test_an_unresolved_role_refuses_before_the_attempt_is_opened",
-   "restoration": "selected-module-boundary",
-   "seams": [
-    "monkeypatch"
-   ]
-  },
-  {
-   "module": "tests.test_dispatch_launch",
-   "owner": "DispatchLaunchTest.test_dispatching_a_sealed_pending_item_readies_it_without_hanging",
-   "restoration": "selected-module-boundary",
-   "seams": [
-    "threads"
-   ]
-  },
-  {
-   "module": "tests.test_dispatch_launch",
    "owner": "LandTest.setUp",
    "restoration": "selected-module-boundary",
    "seams": [
@@ -2318,6 +2277,47 @@
    "restoration": "selected-module-boundary",
    "seams": [
     "environment"
+   ]
+  },
+  {
+   "module": "tests.test_dispatch_launch_command",
+   "owner": "DispatchLaunchTest.established",
+   "restoration": "selected-module-boundary",
+   "seams": [
+    "monkeypatch"
+   ]
+  },
+  {
+   "module": "tests.test_dispatch_launch_command",
+   "owner": "DispatchLaunchTest.setUp",
+   "restoration": "selected-module-boundary",
+   "seams": [
+    "environment",
+    "threads"
+   ]
+  },
+  {
+   "module": "tests.test_dispatch_launch_command",
+   "owner": "DispatchLaunchTest.test_an_unknown_host_refuses_before_the_attempt_is_opened",
+   "restoration": "selected-module-boundary",
+   "seams": [
+    "monkeypatch"
+   ]
+  },
+  {
+   "module": "tests.test_dispatch_launch_command",
+   "owner": "DispatchLaunchTest.test_an_unresolved_role_refuses_before_the_attempt_is_opened",
+   "restoration": "selected-module-boundary",
+   "seams": [
+    "monkeypatch"
+   ]
+  },
+  {
+   "module": "tests.test_dispatch_launch_command",
+   "owner": "DispatchLaunchTest.test_dispatching_a_sealed_pending_item_readies_it_without_hanging",
+   "restoration": "selected-module-boundary",
+   "seams": [
+    "threads"
    ]
   },
   {
